src/requirements.txt
uvicorn
gunicorn
fastapi
pymongo
certifi
langchain
qdrant_client
underthesea
openai
cohere
pandas
python-multipart
pyodbc>=4.0.30
Office365-REST-Python-Client
pydantic
anyio==3.6.2
click==8.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
gunicorn==20.1.0
SQLAlchemy~=2.0.5.post1
